Add 32 bit map tiles and place secondary tileset tile for each connection (tilesets not loaded for now)

+++ include/global.fieldmap.h
@@ -60,7 +60,7 @@ struct BackupMapLayout
 {
     s32 width;
     s32 height;
-    u16 *map;
+    u32 *map;
 };
 
 struct ObjectEventTemplate
